requirements.txt
async-timeout==5.0.1
hiredis==3.1.0
PyMySQL==1.1.0
python-dotenv==1.0.1
redis==5.2.1